renaming script to make more sense

+++ pluto_build/sql/create_rpad_geo.sql
@@ -14,7 +14,7 @@ CREATE TABLE pluto_rpad_geo AS (
 WITH pluto_rpad_rownum AS (
 	SELECT a.*, ROW_NUMBER()
     	OVER (PARTITION BY boro||tb||tl
-      	ORDER BY curavt_act::numeric DESC, land_area::numeric DESC, ease ASC) AS row_number
+      	ORDER BY curavt_act DESC, land_area DESC, ease ASC) AS row_number
   		FROM dof_pts_propmaster a),
 pluto_rpad_sub AS (
 	SELECT *
